backend/requirements.txt
yt-dlp>=2026.01.01
click>=8.1.7
mutagen

# FastAPI stack
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12

# JWT auth
python-jose[cryptography]>=3.3.0

# Password hashing (kept for backward compat with existing werkzeug hashes)
werkzeug>=3.1.0

# HTTP client (stem proxy)
httpx>=0.27.0

# Settings / env
python-dotenv>=1.0.0

# Tests
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
flake8
